ToDo/requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-CKEditor==0.4.6
flask-form==1.0.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
